Home work 7 done
Moved singletone code from home work 5 to header file, added 2
overloaded logContainer() functions to main
